Add :disconnect=>:retry option to Database#transaction, for automatically retrying the transaction on disconnect
This has been requested numerous times over the years.  I think its
a bad idea, but it's possible that it is the best alternative when
dealing with a particularly bad database configuration.

Basically, automatically retrying can cause all sorts of problems.
The most likely way of getting it working is to only retry
at transaction boundries.  For a sane database that rolls back
transactions on disconnect, this should work OK for pure database
code.  However, if you do any non-database work inside the
transaction block, you need to be sure that that the work is
idempotent in addition to making sure that it handles rollbacks
correctly.