Detect an additional disconnect type in the postgres adapter
This handles a bad file descriptor error, and should handle more
types of disconnect errors by using a more general regexp.

While here, also do a better job of handling disconnects, so that
block is not called if a disconnect is detected, since in some
cases block can raise a separate PGError.