Quote channel identifier names when using LISTEN/NOTIFY on PostgreSQL
Before, the identifier names were unquoted, which meant that the
LISTEN/NOTIFY query could break depending on the identifier used.
Quoting them fixes this issue, but note that this makes the
identifier case sensitive, and as such can break backwards
compatibility.  If any of your channels are currently specified
with uppercase characters when the actual channel is lowercase,
or you are doing any manual quoting of the channel identifier,
you need to update your code for these changes.