Fix initializing Database objects on PostgreSQL if there are existing PG_NAMED_TYPES
This got broken when the conversion procs were moved from the native
adapter to the shared adapter.  To fix it, just make sure that the
conversion procs already exist, as they are used themselves when
resetting the conversion procs if there are named types.
Basically, if there are existing named types