Major change in the Database <-> ConnectionPool interface
Historically, Sequel's connection pool has been completely decoupled
from the database.  The database originally just provided a
connection proc to the pool, which the pool called to get a
connection.  The default connection proc just called connect
on the database object.

When I added disconnection handling to the connection pool,
I had to hack it in in an ugly manner, passing the disconnection
proc option via the opts hash.  I'm planning on further improvements
that will require more interaction between the database and
connection pool, and adding another hash entry for a proc is
not a good idea.

Since AFAIK no one uses custom connection and disconnection procs
for the connection pool, change things so that the connection
pool gets a reference to the database object that creates it,
and just call the connect and disconnect_connection methods
directly on the database.  As the disconnect_connection method
was not defined by default, define it as a public method, and
modify the adapters to make it a public method.