Modify the adapter specs to skip checking for specific SQLs in some cases
If the SEQUEL_NO_CHECK_SQLS environment variable is defined,
skip the checks for specific SQLs, only check for correct
behavior.  This is useful if you are using an extension that
modifies the queries that will be sent to the database.

Remove the specific hack in the postgres adapter that already
did this partially for auto parameterization.  Tests with the
auto parameterization extension should set the environment
variable instead.