Fix recursive mutex lock in connection_validator extension
The connection_validator extension was checking connections
for validity while holding the pool mutex.  This greatly
reduces the concurrency and can result in a recursive mutex
lock if the validation itself requires a connection (which
is true on some adapters if the validation SQL was not
cached).

Instead of wrapping next_available, wrap acquire, which is
not called with the mutex lock.  Manually do locking of the
necessary data structures when modifying them.