Make uniqueness validations correctly handle nil values
Previously, the uniqueness validations did not work correctly for
nil values, since they checked if there were other rows where
(column IS NULL).  However, in SQL, NULL = NULL is not true, and
multiple rows are allowed to have NULL values in a column with a
unique constraint.  So skip the uniqueness check if the value is
nil.

If you have a UNIQUE NOT NULL column, you should have a separate
presence check to check for the nil values.