Refactor how dataset methods are added to model datasets
Historically, the way dataset methods were added to datasets was
on a method-by-method basis.  The methods were stored in the
dataset_methods hash, keyed by method name, and each was
individually defined on the dataset's singleton class.

A while back, the dataset_module feature was added, so you
could just extend the dataset with a module, and it would
automatically have all of the module's methods, without having
to define the methods individually.

This commit removes the method-by-method implementation.  Now,
def_dataset_method calls dataset_module internally to define the
method inside the dataset module.

I've also decided to make things a little more consistent by
using a Module subclass, and allowing subset to be called inside
of a dataset_module block, for easier grouping of dataset methods
in one place.  Now Model.subset, just calls subset in the dataset
module.

Finally, adding public methods to a dataset module outside of the
block now adds the appropriate class methods, using a method_added
hook.

The only backwards incompatible change in this commit is the
removal of the Module#dataset_method attr_reader.