Add better description of qualified identifiers/aliased expressions to README
I'm moving the description of qualified tables to the general
section on qualification. The problem with using the :schema__table
example earlier is that we haven't explained what it means, so it
looks like magic.

I noticed the join table example was also using the __ syntax
before the explanation, so I removed that as well.