Reset Model.instance_dataset when extending the model's dataset
Before, if you used Model.dataset_module to add methods to a
model's dataset, those methods weren't used in the model's
instance_dataset. In most cases, that won't matter, since the
instance dataset is only used for updates and refreshes, but
theoretically you should be able to call such methods on the
instance_dataset as well.

This introduces the reset_instance_dataset to derive the
instance dataset from the dataset.  I chose this path instead
of extending the instance dataset directly, as I believe it
will be more reuseable, and it makes it so that the Module's
extended method is not called twice.