Allow Database#get to take an array to return multiple values, similar to map/select_map
This is very similar to the map/select_map change made some time
go.  For array arguments, we need to make sure we have a valid
alias for each array element, then we just use values_at to
retrieve the resulting values.

This could potentially cause backwards compat issues when using
a conditions specifier:

  DB.get([[:a, 1], [:b, 2]])

That type of code now needs to be updated:

  DB.get(Sequel.expr([[:a, 1], [:b, 2]]))

The same was true of the map/select_map change, but I didn't
realize the backwards compat issues back then.  However, as
nobody reported problems, it's unlikely anyone was relying
on the old behavior.