Remove the _*_dataset_helper private method definition
This private method existed because when association support was
originally added to Sequel, instance_exec was not supported, so
the only way to pass arguments to an instance_evaled block was
to define a method for it.  Now that Sequel has dropped Ruby
1.8.6 support, all versions of ruby supported by Sequel support
instance_exec, so just use that instead.

The public AssociationReflection#dataset_helper_method is also
being removed, since it no longer has any use.