Fix Database#copy_into on jdbc/postgres when an exception is raised
The previous implementation has a serious error in that if a ruby
exception was raised, it committed the copy instead of aborting
it.  The reason this wasn't caught by the specs is that it used
an internal transaction.  However, the method is not specced to
use an internal transaction, and the native postgres adapter does
not use an internal transaction, so remove the transaction usage.
In addition to aborting the copy correctly, also correctly wrap
the underlying exception in a DatabaseError.