Move the #meta_def support for Database, Dataset, and Model to the meta_def extension
This is a backwards incompatible change in that the #meta_def
method is no longer available by default on Database, Dataset,
and Model classes and instances.

Sequel originally depended on the metaid gem that added meta_def
and a handful of other methods to Object.  Over the years I removed
usage of most of the other methods, and restricted their definitions
in Sequel to just Database, Dataset, and Model classes and instances.
However, the metadef method was only used in a couple of places
in Sequel, and it really only saves a few characters in those
places, so I decided to remove usage of meta_def completely.

As meta_def was no longer used internally, and it was originally
only used for implementing Sequel, there was no need to load it
by default and have all Sequel users pay the memory penalty for
it.

Since some users may need it, I've decided to move it to an
extension to make it easy to work with old code.

The core and extension specs still use meta_def significantly, so
the extension is loaded automatically by those specs.  Hopefully
those specs can be cleaned up in the future.