Improve association method performance by caching intermediate dataset
Previously, when getting an association dataset, it would always be
derived from the associated class, which can be an expensive
operation due to the number of dataset clones involved.  Additionally,
all association datasets are extended with a module, which resulted
in Object#extend being called at runtime, leading to a performance
decrease as method caches are cleared.

This commit starts caching the non-instance specific dataset in
the association reflection, so that when an instance requests a
an association dataset, much less work is necessary.

This increases association_dataset method performance by about 179%
and association method performance by about 29% on ruby 1.9 in my
testing.  However, that performance increase only affects
associations without a custom :dataset option.  Associations with a
custom dataset option are marginally slower with the new code.

This change should not have backwards compatibility issues, since
it only affects datasets without custom dataset options, and the
standard association dataset just uses filters/joins.  The spec
changes are just for reordering of conditions, since the
:conditions option is now applied to the dataset before the
foreign key filters.

While here, switch from using filter to where, since we always want
to affect the WHERE clause with association conditions.