Allow the :dataset association option to accept the association reflection as an argument
This fixes the limitation of the previous commit, where custom
:dataset options could not benefit from the caching of intermediate
association datasets.  With this change, the :dataset proc can
accept an option, in which case it is passed the association
reflection, on which the associated_dataset method can be called.

This change has some minor backwards incompatibility.  The private
_association_dataset method (note preceding underscore) is no longer
defined, and the public AssociationReflection#_dataset_method method
is also no longer defined.