Increase virtual row performance by using a shared VirtualRow instance
I've thought about doing this for a while, but never got around to
it.  Basically, Sequel::VirtualRow instances are purely used for
DSL purposes, they are stateless.  Instead of creating new
instances everytime you want to evaluate a virtual row, use a
shared instance.  This improves performance up to 100% in the best
case, and about 30% in the typical case for something like {a > 1}.