Remove row_number_column addition in the emulated offset code
Instead of having the row number column be selected, and trying
to filter it out in all of the adapters, add a selection around
the subquery so the row number column is not returned at all.

This greatly simplifies the code, fixes any issues with ambiguous
row number columns in nested queries, and should improve processing
speed for all fetches.  The slight negative is that it may
require a query to get the columns for queries with offsets, and
potentially multiple queries if there are subqueries with offsets.
That can be mitigated by using the columns_introspection extension.

This causes issues with the correlated subquery eager limit
strategy on Microsoft SQL Server, but as that supports window
functions, there is no reason to use correlated subqueries there
(and plenty of adapters don't support that strategy).

While here, fix columns to call naked, since the row_proc shouldn't
be called if you are just retrieving the columns.