Have #last and #paged_each for model datasets order by the model's primary key by default
I originally had planned to do this to #first as well, for
consistency, but it would have required extensive spec changes.
Also, such a consistency is not necessary.  The reason for doing
this for #last and #paged_each is that those methods require
orders, so having a default order by primary key allows them
to work without specifying an order explicitly.  #first already
works without specifying an order explicitly, so it doesn't need
a similar change.