Raise serialization failures/deadlocks as Sequel::SerializationFailure exceptions
This builds on the support added for the ConstraintViolation
exceptions.  Having a separate exception class for these types of
errors is beneficial as these errors can usually be worked around
just by retrying the transaction.

There is no integration test for this support.  Simulating
serialization failures when testing is prone to locking up
the testing program, and the many databases that Sequel supports
handle different types of serialization issues in different ways.