Add the :retry_on and :num_retries transaction options for automatically retrying transactions
This expands the supported added a while back for the
:disconnect=>:retry option, and generalizes it to handle other
types of errors.  For example, a UniqueConstraintViolation or
SerializationFailure exception is a prime candidate for retrying
a transaction.

I've historically been against adding this as I think it's likely
that people put non-idempotent actions in their transaction blocks.
However, the recommended course of action previously has been to
use a loop in your user code, and I think this makes for a much
nicer API.  The warnings about idempotency remain, if the user wants
to ignore them, that is their problem.

In order to avoid infinite loops, the :num_retries option has been
added, and defaults to 5 retries.  With the previous code, an
infinite loop was unlikely as not much code will repeatedly raise
a disconnection error.  However, by allowing the user to pass
an exception class, we don't want to cause an infinite loop just
because they provide a CheckConstraintViolation.  For the people
that really need an infinite loop, allow an explicit nil option
for :num_replies.