Add unlimited_update plugin to work around MySQL warning in replicated environments
I've had a couple requests for this recently, and since the work
around is pretty simple, I think a plugin makes sense.  This could
be done by default, but I don't want to make all applications pay
the cost of the additional dataset clone when the problem only
affects a small portion of users, and even for them it only causes
a warning.