Make json_serializer from_json and array_from_json class methods secure by default
This makes both more similar to the from_json instance method by
default, using set or set_fields and not allowing the setting
of arbitrary columns or associations.  An :associations option
has been added to support specific associations, and the
:all_columns and :all_associations options have been added to get
the previous insecure behavior.

The json_create class method uses the :all_columns and
:all_associations options, so it should be fully backwards
compatible if you change Sequel.parse_json to use
:create_additions=>true when parsing JSON.

The json_create method is now considered deprecated, I will
probably be removing the :all_columns and :all_associations
options as well as the json_create method starting in Sequel 4.