Make xml_serializer plugin secure by default
This contains non-backwards compatible changes to the xml_serializer
plugin, similar to the changes recently made to the json_serializer
plugin, and for the same reason.  The new API/options are very
similar to the json_serializer plugin.  You can get back the old
behavior using the :all_associations and :all_columns options.