Make ilike use UPPER and LIKE instead of ILIKE by default
ILIKE is a PostgreSQL extension, so Sequel shouldn't default to
using it.  All other adapters previously override the handling
for ilike.  Most of them used UPPER in some form, so just make
that the default and DRY up some adapter code.

Remove the like_uses_escape? method added recently, and just
always use ESCAPE.  The only database that didn't support this
was access, so just override the shared access adapter to not
use ESCAPE.

This shouldn't an effect for databases that Sequel currently
supports.  For databases without special support, this should
make ILIKE work by default.