Remove Dataset#def_mutation_method instance method
The instance method creates a singleton class by defining a
method on an instance, which is going to invalidate method
caches in most ruby implementations.  Plus, it's not something
that should be done.

Replace the need for an instance level def_mutation_method by
allowing the class level def_mutation_method to take a :module
option for the module in which to place the defined method.
If you really want the old behavior, you can call the class
method with a :module option set to the singleton class of the
dataset instance.

As def_mutation_method was public, this is a backwards-incompatible
change.