Remove bad model typecasting of money type on PostgreSQL (Fixes #624)
Typecasting money strings to decimal doesn't work, since it
doesn't handle the locale-specific strings.  The dataset
changes were made a while back, but the model typecasting
code was not updated.

Leave the money/smallmoney typecasting on MSSQL.  It appears
to work correctly, and I don't want to risk breaking existing
code that relies on it.