Have associations, composition, serialization, and dirty plugins clear caches in some additional cases
Previously, if the model's dataset supported insert_select, the
caches used by the associations, composition, serialization, and
dirty plugins would not be cleared when saving new records.  Also,
if set_values was used directly, the cache would not be cleared.

This makes it so that set_values clears the cache, and since all
types of refreshes run through set_values, this handles the cases
without other problems.

One issue with this is that set_values is called when creating
model instances.  As such, this has a performance hit when
creating new instances.  This makes Model.call significantly slower,
but actually usage with a database is only about 3% slower when
fetching large numbers of records, which I consider acceptable.