Dataset #select_map, #select_order_map, and #get no longer support a plain string inside an array of arguments
Before, they supported a plain string only by accident, since
recursion was used and they support strings inside of
Identifier and similar expression objects.

Basically, we need to be able to guess the symbol that will be
used in the returned hash, and if a plain string is used, it is
unlikely the database will use that string as the returned
column name.