Minor cleanups in Database.connect
Simplify the connection option code and the connection block
return value code.

Set adapter scheme should always be called with a symbol.  With
the previous code, the ADAPTER_MAP would work if given a string,
but the @scheme would be incorrect.  All internal adapters use
a symbol, and external adapters should as well.