Add Plugins.def_dataset_methods for easily defining class methods that call dataset methods
Previously, only the base model plugin added dataset methods, and
that was mostly to get the core dataset methods, not dataset
methods specific to the base plugin.  It relied on the plugin
behavior of adding class methods for public methods in
DatasetMethods.   That's problematic for 2 reasons.  First, it
ends up creating class methods you don't want.  Second, it
defines the class methods inside the class itself instead of
in ClassMethods in the plugin.