Reflect constraint_validations extension :allow_nil=>true setting in the database constraints
Previously, if :allow_nil was false or not set, it specifically
checked that the column was not NULL.  However, if :allow_nil was
set, it didn't specifically allow NULL in the constraint. This
results in constraints failing for NULL values even if :allow_nil
is set.

Unfortunately, if you previously generated constraints using this
extension with the :allow_nil option set, you need to drop
those constraints and regenerate them.