Remove meta_def extension use from core specs
The core specs should not use extensions, to ensure that nothing
in core depends on an extension.

This is a mostly mechanical change, moving from using the
meta_def Database/Dataset instance method to using a instance
method in the spec example group class.

Note that the meta_def extension is still used in the extension
specs.  However, as the extension specs would load it anyway and
it makes a global change, there seems little point to changing the
them.