Reference driver before loading database-specific code in jdbc adapter (Fixes #646)
If the Database attempts to connect on initialization (which at
least PostgreSQL does in some circumstances), the jdbc adapter
breaks because the driver class is not referenced first, and
java.sql.DriverManager.getConnection cannot find it otherwise
(which is a longstanding bug in JRuby, IMO).  Anyway, making
sure the driver class is referenced first fixes the issue.