Skip validates_unique validation if underlying columns are not valid
If the underlying columns are not valid, the database query to
determine uniqueness could raise an error (for example, an integer
column not in integer format).  That defeats the purpose of
attempting to validate first, so skipping the uniqueness check is
probably best.  The model object is going to be considered invalid
anyway, so the only behavior change other than not raising the
error is that there will be no errors entry for the uniqueness
validation in some cases.