Allow validates_type validation to accept an array of allowable classes
There are some cases where multiple classes would be valid for a
given column.  The most obvious is boolean columns, which can be
either TrueClass or FalseClass.

While here, slightly change the default validates_type message.
Before, this used a capital class name, but that is inconsistent
with the rest of the messages.  Now it uses a lowercase name.