Add validates_schema_types validation for validating column values are instances of an appropriate class
Previously, one of the recommended approaches for dealing with invalid
input with Sequel in web applications was to use the not_string
validation in combination with raise_on_typecast_failure = false.
The idea behind this is that invalid input comes in the form of a
string, so you could use this on non-string types to validate that
the value was typecast correctly.

Unfortunately, while this works in most cases, it is a flawed
approach.  First of all, web application input can also come in the
form of a hash or array (default rack param parsing), and the
not_string validation wouldn't catch that.  Second, you can't use
the not_string validation for string or blob columns, so it isn't
generally applicable.

This new approach uses the database schema information in combination
with the recently added schema_type_class method to get the
appropriate ruby class for a given schema type symbol, and check
that the column value is an instance of that class.

This should be safe to use for all model instances.  Unhandled
database types won't be affected, nor will database types that
have no appropriate ruby class (such as enum).  Every column type
with a handled type class should have a value that is an instance
of the related type, assuming that typecasting works correctly.

The only time this will cause problems is if the database does
not return correctly typecasted values, in which case you'll want
to use the typecast_on_load plugin.