Make Postgres::HStoreOp#- explicitly cast a string argument to text, to avoid PostgreSQL assuming it is an hstore
PostgresSQL has two different - function/operators on hstore,
one for text and one for hstore.  However, if an explicit cast
is not used, the type is unknown, and PostgreSQL gives higher
precedence to the function that accepts an hstore.  As the
string is probably not in hstore format, this usually results
in an "Unexpected end of string" error from the database.