Fix race condition related to prepared_sql for newly prepared statements (Fixes #651)
Previously, the code was racy because the prepared statement SQL
was not finalized before prepare returned.  If two separate
threads then tried to prepare the SQL for the same time statement
and the SQL hadn't been prepared before, the code could break.

Fix this by making sure that the prepared statement SQL is
finalized before prepare returns.  Note that this only needs to
be done if prepare is given a prepared statement name.  If the
name is nil, it is being called by call, which doesn't need to
prepare the SQL first, since it will be working on a thread-local
clone of the dataset.