Fix auto_validations to handle empty values correctly
auto_validations should mirror the database's constraints.  Just
because there is a NOT NULL constraint on the column does not
mean that values such as empty strings are not allowed.  The
constraint_validation extension/plugin can be used to handle
the case where you want to automatically disallow empty strings.