Support the :use_iso_date_format Database option in the postgres adapter, for per-Database specific behavior
Previously, there was only the global use_iso_date_format setting,
which was difficult to change since the adapter had to be loaded
first.

Keep the global setting as a default, but allow modification
per Database via an option.  It doesn't make sense to make this
modifiable after initialization, since that could result in
some connections using ISO and some not.

Note that this moves the use_iso_date_setting out of the pg_types
support.  This is because other drivers can load the pg_types
support and they shouldn't access that setting.  This does make
the default date parsing in pg_types use Date.parse instead of
the more optimized ISO-specific code, but these days I'm not sure
that matters.