Remove some bad code in Sequel::Postgres module RDoc
This code doesn't work correctly if Sequel connects on
initialization, such as when using the pg_hstore extension.
In general, we should not be encouraging using the global
settings, so just remove much of the text and tell people to
use the options.

The postgres-adapter specific use_iso_date_format method
is mentioned in the connecting to a database guide, so
there is no need to mention it in the shared adapter module RDoc.