Support a :search_path Database option for setting PostgreSQL search_path
This is a much better way to set a default schema to use than
Sequel's default_schema setting (which will be going away in
Sequel 4).

The quoting of schema names is kind of ugly, but I don't want
to use the standard quoting methods in Dataset, since they
may require a connection (leading to an infinite loop).