Handle domain types in composite types in the pg_row extension
The way PostgreSQL works with domain types, is that on retrieval,
the oid returned to the client is the oid of the underlying type,
not of the base type itself.  This is why if you just select a
domain type, it is typecast correctly on retrieval (assuming the
underlying base type can be typecast correctly).

The pg_row extension didn't handle domain types specially
previously, just doing a check for the attribute types.  So for
domain types, it would get the domain type oid, which Sequel
probably doesn't have a conversion proc for.  This resulted in
domain types being returned as strings.

This change looks at the underlying type of the domain type, and
using its oid if available, resulting in domain types being
typecast correctly.

Note that Sequel's schema parsing does not do special handling for
domain types, so the schema :oid values and db_type reflect the
domain type and not the underlying type.  I'm not sure if that
should be changed or not.