Make some currently global extensions loadable by Database/Dataset#extension
In Sequel 4, these extensions will no longer be global extensions,
and you will have to load them with Database/Dataset#extension to
get the desired behavior.  This adds the necessary no-op callbacks
for forwards compatibility.

This updates bin/sequel as well so it won't need to be changed in
Sequel 4.