Don't reset conversion procs when loading pg_* extensions
Previously, for the extensions that add entries to PG_TYPES, Sequel
reset the conversion_procs as the path of least resistence.  However,
that can lead to extra work, and may lose custom conversion procs
added manually.

Change the implementation so that just the procs the extension
adds are added to the database when the database loads the support.