Deprecate core_extensions if not explicitly loaded
This adds the main deprecation in Sequel 4, deprecating all of
the core_extensions methods unless they have been explicitly
loaded.  This is done by making a copy of the core_extensions,
and then adding deprecation warnings to all of the methods it
adds.  If the core_extensions are loaded, its implementation will
override the default deprecated one, so the deprecation messages
will go away.

This also adds the ruby18_symbol_extensions extension for the
Symbol#{[],<,<=,>,>=} methods, which Sequel previously overrode
by default on ruby 1.8.  I really want to discourage their use,
so they are deprecated by default even if the core_extensions
are loaded explicitly.  You have to load the new extension if you
want the behavior.

While working on this, I noticed some of the core extensions
were not being tested, because the related methods were in an
included module.  So add specs for those.

This modifies the core_extensions_spec to handle different types
of environments via the SEQUEL_DEPRECATED_CORE_EXTENSIONS
environment variable:

* missing: runs with just the explicit core extensions and with the
  ruby18 symbol extensions on ruby 1.8
* OVERRIDDEN: runs with the deprecated core extensions and the
  explicit core extensions, without the ruby18 symbol extensions
* SYMBOL18: runs with just the deprecated core extensions and the
  the ruby18 symbol extensions
* OVERRIDDENSYMBOL18: runs with the deprecated core extensions, the
  explicit core extensions, and the ruby18 symbol extensions
* other value: runs with just the the deprecated core extensions