Deprecate automatic reciprocals in certain cases
If the automatic reciprocal is ambiguous, or if the only candidates
for the reciprocal have conditions or a block or a different
primary key (or the current association has conditions or a block),
at best the correct answer will only be chosen 50% of the time.
It's best to force the user to make a choose of reciprocal in that
case.

While here, refactor the reciprocal determination so that
many_to_many does not have a different implementation.  Also, set
up explicit reciprocals in the tree plugin so that the rcte_tree
plugin doesn't end up with ambiguous ones.