Deprecate nil dataset options and modifying initial dataset
There isn't any reason to support nil as an argument, since the
default argument should be a hash.

The deprecation of modifying the initialize dataset is a
performance optimization, since it means that in Sequel 4, the
default dataset options hash can be a shared frozen hash.  As
clone is used for dataset modifications, none of the current specs
do this.

In the mutation methods, instead of modifying the @opts hash, just
reassign it to the new hash.  This should be faster, avoids
the future issues with a shared frozen hash, and is even more
correct if the opts for the clone have had entries deleted.