Don't deprecate reciprocals if current association has conditions/block
If the current association has conditions or a block, but the
reciprocal association does not, it should be safe to use
the reciprocal.  Only if the reciprocal association has a
condition or block should it not be safe. Example:

  Album.many_to_one :artist
  Artist.one_to_many :a_albums, :conditions=>{:name=>/^a/i},
    :class=>Album

Here, artist should not pick a_albums as a reciprocal, since the
album name may not start with "a".  However, a_albums should have
artist as the reciprocal, since any matching album should have
that artist.